requirements.txt
Pillow==8.3.2
qrcode==7.3.1